Split safety model parity gates
